requirements.txt
numba==0.56.4
Fiona==1.9.2
GeoAlchemy2==0.13.2
geopandas==0.12.0
psycopg2-binary==2.9.6
pyproj==3.5.0
pyshp==2.3.1
python-dotenv==0.21.0
Shapely==1.8.5.post1
SQLAlchemy==2.0.9
watchdog==3.0.0
wheel==0.37.1

